Move bundle accessor into seperate file
